Switch order of ScrollModeProvider and RowsAndColumnsProvider
